Move some more code around
