+++ miden-tx/tests/integration/scripts/p2idr.rs
@@ -12,7 +12,7 @@ use miden_objects::{
     assets::{Asset, AssetVault, FungibleAsset},
     crypto::rand::RpoRandomCoin,
     notes::NoteType,
-    testing::storage::DEFAULT_AUTH_SCRIPT,
+    testing::account_code::DEFAULT_AUTH_SCRIPT,
     transaction::TransactionArgs,
     Felt,
 };
