+++ miden-tx/tests/integration/wallet/mod.rs
@@ -11,7 +11,7 @@ use miden_objects::{
     assets::{Asset, AssetVault, FungibleAsset},
     crypto::dsa::rpo_falcon512::SecretKey,
     notes::{NoteTag, NoteType},
-    testing::storage::{prepare_word, DEFAULT_AUTH_SCRIPT},
+    testing::{account_code::DEFAULT_AUTH_SCRIPT, prepare_word},
     transaction::TransactionArgs,
     Felt, Word, ONE, ZERO,
 };
@@ -136,9 +136,8 @@ fn prove_send_asset_via_wallet() {
 
     assert_eq!(tag.validate(note_type), Ok(tag));
 
-    let tx_script_code = ProgramAst::parse(
-        format!(
-            "
+    let var_name = &format!(
+        "
         use.miden::contracts::auth::basic->auth_tx
         use.miden::contracts::wallets::basic->wallet
 
@@ -152,14 +151,12 @@ fn prove_send_asset_via_wallet() {
             call.auth_tx::auth_tx_rpo_falcon512
         end
         ",
-            recipient = prepare_word(&recipient),
-            note_type = note_type as u8,
-            tag = tag,
-            asset = prepare_word(&fungible_asset_1.into())
-        )
-        .as_str(),
-    )
-    .unwrap();
+        recipient = prepare_word(&recipient),
+        note_type = note_type as u8,
+        tag = tag,
+        asset = prepare_word(&fungible_asset_1.into())
+    );
+    let tx_script_code = ProgramAst::parse(var_name.as_str()).unwrap();
     let tx_script = executor.compile_tx_script(tx_script_code, vec![], vec![]).unwrap();
     let tx_args: TransactionArgs = TransactionArgs::with_tx_script(tx_script);
