+++ objects/src/testing/account.rs
@@ -4,17 +4,29 @@ use alloc::{
 };
 use core::fmt::Display;
 
-use assembly::{ast::ModuleAst, Assembler};
+use assembly::Assembler;
 use miden_crypto::merkle::MerkleError;
 use rand::Rng;
 
 use super::{
+    account_code::{mock_account_code, DEFAULT_ACCOUNT_CODE},
     account_id::{str_to_account_code, AccountIdBuilder},
-    storage::{AccountStorageBuilder, DEFAULT_ACCOUNT_CODE},
+    assets::non_fungible_asset,
+    constants::FUNGIBLE_ASSET_AMOUNT,
+    storage::{
+        generate_account_seed, storage_item_0, storage_item_1, storage_item_2, storage_map_2,
+        AccountSeedType, AccountStorageBuilder,
+    },
 };
 use crate::{
-    accounts::{Account, AccountCode, AccountStorage, AccountStorageType, AccountType, SlotItem},
-    assets::{Asset, AssetVault},
+    accounts::{
+        account_id::testing::{
+            ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN, ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_1,
+            ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_2, ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN,
+        },
+        Account, AccountCode, AccountId, AccountStorage, AccountStorageType, AccountType, SlotItem,
+    },
+    assets::{Asset, AssetVault, FungibleAsset},
     AccountError, AssetVaultError, Felt, Word, ZERO,
 };
 
@@ -159,23 +171,72 @@ impl Display for AccountBuilderError {
 #[cfg(feature = "std")]
 impl std::error::Error for AccountBuilderError {}
 
-// TESTING
+// MOCK ACCOUNT
 // ================================================================================================
 
-pub const CODE: &str = "
-        export.foo
-            push.1 push.2 mul
-        end
+#[derive(Debug, PartialEq)]
+pub enum MockAccountType {
+    StandardNew,
+    StandardExisting,
+    FungibleFaucet {
+        acct_id: u64,
+        nonce: Felt,
+        empty_reserved_slot: bool,
+    },
+    NonFungibleFaucet {
+        acct_id: u64,
+        nonce: Felt,
+        empty_reserved_slot: bool,
+    },
+}
 
-        export.bar
-            push.1 push.2 add
-        end
-    ";
+pub fn mock_new_account(assembler: &Assembler) -> Account {
+    let (acct_id, _account_seed) =
+        generate_account_seed(AccountSeedType::RegularAccountUpdatableCodeOffChain, assembler);
+    let account_storage = mock_account_storage();
+    let account_code = mock_account_code(assembler);
+    Account::from_parts(acct_id, AssetVault::default(), account_storage, account_code, ZERO)
+}
 
-pub fn make_account_code() -> AccountCode {
-    let mut module = ModuleAst::parse(CODE).unwrap();
-    // clears are needed since they're not serialized for account code
-    module.clear_imports();
-    module.clear_locations();
-    AccountCode::new(module, &Assembler::default()).unwrap()
+pub fn mock_account(account_id: u64, nonce: Felt, account_code: AccountCode) -> Account {
+    let account_storage = mock_account_storage();
+    let account_vault = mock_account_vault();
+    let account_id = AccountId::try_from(account_id).unwrap();
+    Account::from_parts(account_id, account_vault, account_storage, account_code, nonce)
+}
+
+/// Creates an [AssetVault] with 4 assets.
+///
+/// The ids of the assets added to the vault are defined by the following constants:
+///
+/// - ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN
+/// - ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_1
+/// - ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_2
+/// - ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN
+///
+pub fn mock_account_vault() -> AssetVault {
+    let faucet_id: AccountId = ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN.try_into().unwrap();
+    let fungible_asset =
+        Asset::Fungible(FungibleAsset::new(faucet_id, FUNGIBLE_ASSET_AMOUNT).unwrap());
+
+    let faucet_id_1: AccountId = ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_1.try_into().unwrap();
+    let fungible_asset_1 =
+        Asset::Fungible(FungibleAsset::new(faucet_id_1, FUNGIBLE_ASSET_AMOUNT).unwrap());
+
+    let faucet_id_2: AccountId = ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_2.try_into().unwrap();
+    let fungible_asset_2 =
+        Asset::Fungible(FungibleAsset::new(faucet_id_2, FUNGIBLE_ASSET_AMOUNT).unwrap());
+
+    let non_fungible_asset = non_fungible_asset(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN);
+    AssetVault::new(&[fungible_asset, fungible_asset_1, fungible_asset_2, non_fungible_asset])
+        .unwrap()
+}
+
+pub fn mock_account_storage() -> AccountStorage {
+    // create account storage
+    AccountStorage::new(
+        vec![storage_item_0(), storage_item_1(), storage_item_2()],
+        vec![storage_map_2()],
+    )
+    .unwrap()
 }
